lib/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secrets-sdk"
version = "0.1.0"
description = "Python SDK for developer identity and secrets management — Vault, SOPS, config validation, rotation."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Brush Cyber", email = "info@brushcyber.com" },
]
keywords = ["vault", "secrets", "sops", "security", "identity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "hvac>=2.1.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]

[project.scripts]
secrets-sdk = "secrets_sdk.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["secrets_sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
